refactor: simplify imports with index in Tests
