+++ tests/infra/db/mongodb/static/authentication/static-account-mongo-repository.spec.ts
@@ -1,17 +1,16 @@
 import { faker } from '@faker-js/faker'
 import { Collection } from 'mongodb'
-import { AccountMongoRepository } from '@/infra/db/mongodb/account-mongo-repository'
-import { MongoHelper } from '@/infra/db/mongodb/mongo-helper'
-import { mockAddAccountInput } from '@/tests/domain/mocks/mock-account'
-import { throwError } from '@/tests/domain/mocks/test-helper'
+import { mockAddAccountInput, throwError } from '@/tests/domain/mocks'
+import { StaticAccountMongoRepository } from '@/infra/db/mongodb/static/authentication'
+import { MongoHelper } from '@/infra/db/mongodb'
 
 let accountCollection: Collection
 
-const makeSut = (): AccountMongoRepository => {
-  return new AccountMongoRepository()
+const makeSut = (): StaticAccountMongoRepository => {
+  return new StaticAccountMongoRepository()
 }
 
-describe('AccountMongoRepository', () => {
+describe('StaticAccountMongoRepository', () => {
   beforeAll(async() => {
     await MongoHelper.connect(process.env.MONGO_URL as string)
   })
@@ -23,22 +22,6 @@ describe('AccountMongoRepository', () => {
   beforeEach(async() => {
     accountCollection = MongoHelper.getCollection('account')
     await accountCollection.deleteMany({})
-  })
-
-  describe('add', () => {
-    test('Should throw if mongo throws', async() => {
-      const sut = makeSut()
-      jest.spyOn(Collection.prototype, 'insertOne').mockImplementationOnce(throwError)
-      const promise = sut.add(mockAddAccountInput())
-      await expect(promise).rejects.toThrow()
-    })
-
-    test('Should return true on success', async() => {
-      const sut = makeSut()
-      const addAccountInput = mockAddAccountInput()
-      const isValid = await sut.add(addAccountInput)
-      expect(isValid).toBe(true)
-    })
   })
 
   describe('checkByEmail', () => {
