+++ tests/infra/validators/email-validator-adapter.spec.ts
@@ -1,5 +1,5 @@
 import validator from 'validator'
-import { EmailValidatorAdapter } from '@/infra/validators/email-validator-adapter'
+import { EmailValidatorAdapter } from '@/infra/validators'
 
 jest.mock('validator', () => ({
   isEmail(): boolean {
