+++ tests/presentation/mocks/index.ts
@@ -1,0 +1,2 @@
+export * from './mock-authentication'
+export * from './mock-validation'
