+++ tests/presentation/mocks/mock-validation.ts
@@ -1,4 +1,4 @@
-import { type Validation } from '@/presentation/protocols/validation'
+import { type Validation } from '@/presentation/protocols'
 
 export class ValidationSpy implements Validation {
   public input: any
